REF: move common arithmetic ops from Integer/FloatingArray to NumericArray (#38290)
